# backend_monbrobroli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/qualifications/{id}": {
            "delete": {
                "tags": [
                    "Qualifications"
                ],
                "summary": "Supprimer une qualification par ID",
                "description": "Supprime la qualification identifiée par {id}. Retourne 404 si non trouvée.",
                "operationId": "deleteQualification",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la qualification à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Qualification supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Qualification supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Qualification non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Qualification non trouvée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié / token manquant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addQualifications": {
            "post": {
                "tags": [
                    "Qualifications"
                ],
                "summary": "Ajouter des qualifications pour l'étudiant connecté",
                "description": "Cette API permet à l'étudiant authentifié d'ajouter une ou plusieurs qualifications.",
                "operationId": "addQualifications",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "qualifications": {
                                        "description": "Liste des qualifications à ajouter",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Licence en Informatique"
                                                },
                                                "date_debut": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2020-09-01"
                                                },
                                                "date_fin": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-06-30"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Qualification ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Qualification ajoutée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/updateQualification": {
            "post": {
                "tags": [
                    "Qualifications"
                ],
                "summary": "Mettre à jour les qualifications d’un étudiant",
                "operationId": "67dfc868f3f4fbadb8b1dd319df2e4ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qualifications"
                                ],
                                "properties": {
                                    "qualifications": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "objet",
                                                "detail",
                                                "date_debut",
                                                "date_fin"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "objet": {
                                                    "type": "string",
                                                    "example": "Développement Web"
                                                },
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Laravel & Vue.js"
                                                },
                                                "date_debut": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-01-01"
                                                },
                                                "date_fin": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-01-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Qualification mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La qualification a été modifiée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}